split index.ts from app definitions
